refactor:  alteracoes de altura e largura para card de eventos e add slug para comunidades
